+++ src/scicloj/ml/smile/classification.clj
@@ -455,8 +455,10 @@
 (defn- predict
   [feature-ds thawed-model {:keys [target-columns
                                    target-categorical-maps
+                                   target-datatypes
                                    options
-                                   model-data]}]
+                                   model-data
+                                   ]}]
   ;; (errors/when-not-error target-categorical-maps "target-categorical-maps not found. Target column need to be categorical.")
   (let [n-labels (model-data :n-labels)
         entry-metadata (model-type->classification-model
@@ -473,7 +475,7 @@
                                            n-labels
                                            target-categorical-maps))
         mapped-predictions
-        (-> (ds-mod/probability-distributions->label-column finalised-predictions target-colname)
+        (-> (ds-mod/probability-distributions->label-column finalised-predictions target-colname (get target-datatypes target-colname))
             (ds/update-column target-colname
                               #(vary-meta % assoc :column-type :prediction)))]
     mapped-predictions))